Rails upgrade 4.2 to 5.2
- updated add index migration
- removed thredded upgrate migrations

+++ db/migrate/20190627071643_drop_thredded_messageboard_users_table.rb
@@ -1,6 +1,6 @@
 class DropThreddedMessageboardUsersTable < ActiveRecord::Migration[5.2]
   def up
-    drop_table :thredded_messageboard_users
+    drop_table :thredded_messageboard_users if table_exists? :thredded_messageboard_users
   end
 
   def down
